CU-3bf4jg1 - Remove Console logs
