+++ package.json
@@ -16,7 +16,7 @@
     "@fortawesome/free-solid-svg-icons": "^6.1.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@mui/material": "^5.9.3",
-    "@nabcellent/sui-react": "^2.2.3",
+    "@nabcellent/sui-react": "^2.4.13",
     "@reduxjs/toolkit": "^1.8.3",
     "echarts": "^5.3.3",
     "formik": "^2.2.9",
@@ -33,6 +33,7 @@
     "yup": "^0.32.11"
   },
   "devDependencies": {
+    "@types/node": "^18.7.16",
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@types/react-router-dom": "^5.3.3",
