+++ src/pages/dashboards/default/RecentPersonalTransactions.tsx
@@ -1,6 +1,7 @@
 import { ComponentLoader, SectionError } from '@nabcellent/sui-react';
 import { useGetRecentPersonalAccountTransactionQuery } from 'features/savings/savingsApi';
 import SavingsTransactionTable from 'components/tables/SavingsTransactionTable';
+import { logger } from 'utils/logger';
 
 const RecentPersonalTransactions = () => {
     let { data: transactions, isLoading, isSuccess, isError, error } = useGetRecentPersonalAccountTransactionQuery();
@@ -8,7 +9,7 @@ const RecentPersonalTransactions = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !transactions) return <ComponentLoader/>;
 
-    console.log(transactions);
+    logger.log(transactions);
 
     return <SavingsTransactionTable title={'Recent Personal Transactions'} transactions={transactions}
                                     entity={'personal-accounts'}/>;
