+++ src/pages/group-accounts/Transactions.tsx
@@ -1,6 +1,7 @@
 import { SectionError, SectionLoader } from '@nabcellent/sui-react';
 import SavingsTransactionTable from 'components/tables/SavingsTransactionTable';
 import { useGetGroupAccountTransactionsQuery } from 'features/group-accounts/groupAccountApi';
+import { logger } from 'utils/logger';
 
 const Transactions = () => {
     let { data: transactions, isLoading, isSuccess, isError, error } = useGetGroupAccountTransactionsQuery();
@@ -8,7 +9,7 @@ const Transactions = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !transactions) return <SectionLoader/>;
 
-    console.log(transactions);
+    logger.log(transactions);
 
     return <SavingsTransactionTable title={'Group Account Transactions'} transactions={transactions}
                                     entity={'group-accounts'}/>;
