+++ src/pages/group-accounts/index.tsx
@@ -6,6 +6,7 @@ import SidoohAccount from 'components/SidoohAccount';
 import { currencyFormat, DataTable, SectionError, SectionLoader, StatusChip, TableDate } from '@nabcellent/sui-react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faEye } from '@fortawesome/free-regular-svg-icons';
+import { logger } from 'utils/logger';
 
 const Index = () => {
     let {data: accounts, isLoading, isSuccess, isError, error} = useGetGroupAccountsQuery();
@@ -13,7 +14,7 @@ const Index = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !accounts) return <SectionLoader/>;
 
-    console.log(accounts);
+    logger.log(accounts);
 
     return (
         <Card className={'mb-3'}>
