+++ src/pages/groups/CollectiveInvestments.tsx
@@ -1,6 +1,7 @@
 import { SectionError, SectionLoader } from '@nabcellent/sui-react';
 import CollectiveInvestmentsTable from 'components/tables/CollectiveInvestmentsTable';
 import { useGetGroupCollectiveInvestmentsQuery } from 'features/groups/groupsApi';
+import { logger } from 'utils/logger';
 
 const PersonalCollectiveInvestments = () => {
     let {data: investments, isLoading, isSuccess, isError, error} = useGetGroupCollectiveInvestmentsQuery();
@@ -8,7 +9,7 @@ const PersonalCollectiveInvestments = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !investments) return <SectionLoader/>;
 
-    console.log(investments);
+    logger.log(investments);
 
     return <CollectiveInvestmentsTable title={'Group Collective Investments'} investments={investments}/>;
 };
