+++ src/pages/personal-accounts/CollectiveInvestments.tsx
@@ -1,6 +1,7 @@
 import { useGetPersonalCollectiveInvestmentsQuery } from 'features/personal-accounts/personalAccountApi';
 import { SectionError, SectionLoader } from '@nabcellent/sui-react';
 import CollectiveInvestmentsTable from 'components/tables/CollectiveInvestmentsTable';
+import { logger } from 'utils/logger';
 
 const PersonalCollectiveInvestments = () => {
     let {data: investments, isLoading, isSuccess, isError, error} = useGetPersonalCollectiveInvestmentsQuery();
@@ -8,7 +9,7 @@ const PersonalCollectiveInvestments = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !investments) return <SectionLoader/>;
 
-    console.log(investments);
+    logger.log(investments);
 
     return <CollectiveInvestmentsTable title={'Personal Account Collective Investments'} investments={investments}/>;
 };
