+++ src/pages/personal-accounts/Transactions.tsx
@@ -3,6 +3,7 @@ import {
 } from 'features/personal-account-transactions/personalAccountTransactionApi';
 import { SectionError, SectionLoader } from '@nabcellent/sui-react';
 import SavingsTransactionTable from 'components/tables/SavingsTransactionTable';
+import { logger } from 'utils/logger';
 
 const Transactions = () => {
     let { data: transactions, isLoading, isSuccess, isError, error } = useGetPersonalAccountTransactionsQuery();
@@ -10,7 +11,7 @@ const Transactions = () => {
     if (isError) return <SectionError error={error}/>;
     if (isLoading || !isSuccess || !transactions) return <SectionLoader/>;
 
-    console.log(transactions);
+    logger.log(transactions);
 
     return <SavingsTransactionTable title={'Personal Account Transactions'} transactions={transactions}
                                     entity={'personal-accounts'}/>;
