+++ src/vite-env.d.ts
@@ -1,5 +1,7 @@
 /// <reference types="vite/client" />
 
+import { LogLevel } from '@nabcellent/sui-react';
+
 interface ImportMetaEnv {
     readonly VITE_ACCOUNTS_API_URL: string
     readonly VITE_SAVINGS_API_URL: string
@@ -9,6 +11,8 @@ interface ImportMetaEnv {
     readonly VITE_PAYMENTS_DASHBOARD_URL: string
     readonly VITE_NOTIFY_DASHBOARD_URL: string
     readonly VITE_USSD_DASHBOARD_URL: string
+
+    readonly VITE_LOG_LEVEL: LogLevel
 }
 
 interface ImportMeta {
